+++ package.json
@@ -19,8 +19,6 @@
     "@mui/lab": "^5.0.0-alpha.83",
     "@mui/material": "^5.8.1",
     "hangul-js": "^0.2.6",
-    "mobx": "^6.3.12",
-    "mobx-react": "^7.2.1",
     "next": "12.1.5",
     "next-i18next": "^11.0.0",
     "react": "18.1.0",
