--- src/stores/RootStore.ts
@@ -1,9 +1,0 @@
-import { AppStore } from './app/AppStore';
-
-export class RootStore {
-    app: AppStore;
-
-    constructor() {
-        this.app = new AppStore();
-    }
-}
